implement package configsecret - switch delimiter to "_"
