+++ railgun/pkg/configsecret/common.go
@@ -1,0 +1,3 @@
+package configsecret
+
+const KeyDelimiter = "_"
